pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blockdataapis"
version = "1.0.2"
description = "Reads blockchain data from apis gotten from exchanges, NFT Marketplaces, explorers etc"
readme = "README.md"
authors = [{ name = "Aminat Amusa", email = "minaomobonike@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["Blockchain", "Cryptocurrencies", "NFTs"]
dependencies = [
    "requests>=2.28.1",
    "python-decouple>=3.6",
    "pytest>=3.0.3",

]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pandas"]

[project.urls]
Homepage = "https://github.com/minat-hub/Blockdataapis"

[project.scripts]
blockdataapis = "blockapiwrapper.base"



[tool.bumpver]
current_version = "2022.1004-alpha"
version_pattern = "YYYY.BUILD-TAG"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]
"setup.py" = [
    "{version}",
    "{pep440_version}",
]
"README.md" = [
    "{version}",
    "{pep440_version}",
]

